Crate wirepact_translator

Structs

CheckRequest

EgressResult

Struct that contains the egress result for the Translator::egress method. Used by the respective constructors to signal a specific result to Envoy.

IngressResult

Struct that contains the ingress result for the Translator::ingress method. Used by the respective constructors to signal a specific result to Envoy.

Status

A gRPC status describing the result of an RPC call.

TranslatorConfig

Basic translator configuration. Used in run_translator to configure and run the translator and the Pki.

Constants

HTTP_AUTHORIZATION_HEADER

Name of the default HTTP authorization header.

WIREPACT_IDENTITY_HEADER

Name of the custom HTTP header that is used by WirePact to identify the signed JWT.

Traits

Translator

Translator for ingress and egress communication. This trait is used in the respective gRPC servers to translate authentication data into a signed JWT and vice versa.

Functions

run_translator

Runs the Translator and the Pki in a separate thread. The ingress/egress gRPC servers are configured with their respective ports. To stop the translator, the threads listen for SIGINT and SIGTERM (or ctrl_c in Windows).

Attribute Macros

async_trait