Trait wirepact_translator::Translator

pub trait Translator: Send + Sync {
    fn ingress<'life0, 'life1, 'life2, 'async_trait>(
        &'life0 self, 
        subject_id: &'life1 str, 
        request: &'life2 CheckRequest
    ) -> Pin<Box<dyn Future<Output = Result<IngressResult, Status>> + Send + 'async_trait>>
    where
        'life0: 'async_trait,
        'life1: 'async_trait,
        'life2: 'async_trait,
        Self: 'async_trait;
    fn egress<'life0, 'life1, 'async_trait>(
        &'life0 self, 
        request: &'life1 CheckRequest
    ) -> Pin<Box<dyn Future<Output = Result<EgressResult, Status>> + Send + 'async_trait>>
    where
        'life0: 'async_trait,
        'life1: 'async_trait,
        Self: 'async_trait;

    fn get_header(
        &self, 
        request: &CheckRequest, 
        header_name: &str
    ) -> Result<Option<String>, Status> { ... }
}

Translator for ingress and egress communication. This trait is used in the respective gRPC servers to translate authentication data into a signed JWT and vice versa.

Required Methods

fn ingress<'life0, 'life1, 'life2, 'async_trait>(
&'life0 self,
subject_id: &'life1 str,
request: &'life2 CheckRequest
) -> Pin<Box<dyn Future<Output = Result<IngressResult, Status>> + Send + 'async_trait>> where
'life0: 'async_trait,
'life1: 'async_trait,
'life2: 'async_trait,
Self: 'async_trait,

Inbound communication translator. Used to transform a signed JWT (if any) into the corresponding authentication data. If no JWT is available or no subject_id can be parsed, the method will not be called and the IngressResult will be IngressResult::forbidden.

fn egress<'life0, 'life1, 'async_trait>(
&'life0 self,
request: &'life1 CheckRequest
) -> Pin<Box<dyn Future<Output = Result<EgressResult, Status>> + Send + 'async_trait>> where
'life0: 'async_trait,
'life1: 'async_trait,
Self: 'async_trait,

Outbound communication translator. Used to transform the authentication data into a signed JWT that contains the users ID.

Provided Methods

fn get_header(
&self,
request: &CheckRequest,
header_name: &str
) -> Result<Option<String>, Status>